+++ lib/src/emoji_picker_utils.dart
@@ -25,13 +25,12 @@ class EmojiPickerUtils {
   RegExp? _emojiRegExp;
 
   /// Returns list of recently used emoji from cache
-  Future<List<RecentEmoji>> getRecentEmojis() async {
-    return EmojiPickerInternalUtils().getRecentEmojis();
-  }
+  Future<List<RecentEmoji>> getRecentEmojis() =>
+      Future.value(EmojiPickerInternalUtils().getRecentEmojis());
 
   /// Filters out emojis not supported on the platform
   Future<List<CategoryEmoji>> filterUnsupported(List<CategoryEmoji> data) =>
-      EmojiPickerInternalUtils().filterUnsupported(data);
+      Future.value(EmojiPickerInternalUtils().filterUnsupported(data));
 
   /// Search for related emoticons based on keywords
   Future<List<Emoji>> searchEmoji(